skip package check if snapshot doesnt exist

+++ build_runner/lib/src/build_script_generate/bootstrap.dart
@@ -131,15 +131,16 @@ Future<int> generateAndRun(List<String> args, {Logger logger}) async {
 Future<int> _createSnapshotIfNeeded(Logger logger) async {
   var assetGraphFile = File(assetGraphPathFor(scriptSnapshotLocation));
   var snapshotFile = File(scriptSnapshotLocation);
+  var snapshotExists = await snapshotFile.exists();
 
   // If we failed to serialize an asset graph for the snapshot, then we don't
   // want to re-use it because we can't check if it is up to date.
-  if (!await assetGraphFile.exists() && await snapshotFile.exists()) {
+  if (snapshotExists && !await assetGraphFile.exists()) {
     await snapshotFile.delete();
     logger.warning('Deleted previous snapshot due to missing asset graph.');
   }
 
-  if (!await _checkImportantPackageDeps()) {
+  if (snapshotExists && !await _checkImportantPackageDeps()) {
     await snapshotFile.delete();
     logger.warning('Deleted previous snapshot due to core package update');
   }
